tools/cloud-build/requirements.txt
pre-commit
ansible
pyyaml
